Add test_badge_should_have_entity_id test
Note that during refactoring, the path forward appears to have many @property
statements for each of the attributes. This feels like a code smell and
something that should be done dynamically.

    @property
    def entity_id(self):
        return self._attrs['entity_id']

This appeared to naturally fit a namedtuple. This approach was strongly
considered when building the Badge class. The reason it was abandoned was a
lack of understanding/experience on how to add custom methods to the resultant
class.

In true TDD fashion, the large entity list will be added with tests. Then, when
a clear path forward is seen, there will be enough test coverage for a good
refactor.