Add test_cli_subcommand_with_token test
Note this required a refactor to delay use of token file until API calls
are made that use it. This lazier approach allows the use of --help in
subcommands without having a functioning token file.

The logic is that one should be able to get the --help before the token
file requirements are enforced.